// web/solicitudes-web/celesol-backend/src/modules/socios/application/dtos/UpdateSocio.dto.ts
export type UpdateSocioDto = {
  apellido?: string;
  celular?: string;
  cuit?: string;
  domicilioCalle?: string;
  domicilioCodigoPostal?: string;
  domicilioLocalidad?: string;
  domicilioNroPuerta?: string;
  email?: string;
  fechaDeNacimiento?: string;
  nombre?: string;
  nroDocumento?: string;
  razonSocial?: string;
  sexo?: string;
  tipoDocumento?: string;
};